requirements-preprocessor.txt
# PDF Processing dependencies
pdf2image==1.16.3
PyMuPDF==1.23.0
PyPDF2==3.0.0
pdfminer.six==20221105
Pillow==10.0.0

# AWS
boto3==1.34.0
botocore==1.34.0

# Utilities
asyncio-throttle==1.0.2 